requirements.txt
# Project uses Python's built-in secrets module for random generation 
Flask==3.0.0
# Project uses Python's built-in secrets module for random generation 
gunicorn==21.2.0
Werkzeug==3.0.1
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
limits==3.9.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0 